utils: move dependency management tools into linters/dependencymgmt/depmanagers
